docs(readme): improve install section (#186)
Install shipit-cli globally instead of dev-dependency to be able to use it from the command line. Without installing it globally, `shipit` cannot be used as command line program.

shipit-deploy needs to be manually installed in the project for the examples below to work, else shipit will throw an error because it cannot find shipit-deploy.